Moving the base task to the project root (#2702)
* Moving the base task to the project root

* Moving the base class more

* Changing the import again
