+++ pokemongo_bot/__init__.py
@@ -15,6 +15,7 @@ from pgoapi import PGoApi
 from pgoapi.utilities import f2i, get_cell_ids
 
 import cell_workers
+from base_task import BaseTask
 from plugin_loader import PluginLoader
 from api_wrapper import ApiWrapper
 from cell_workers.utils import distance
