+++ pokemongo_bot/base_task.py
