+++ pokemongo_bot/cell_workers/catch_lured_pokemon.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from pokemongo_bot.cell_workers.utils import fort_details
 from pokemongo_bot.cell_workers.pokemon_catch_worker import PokemonCatchWorker
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 
 class CatchLuredPokemon(BaseTask):
