+++ pokemongo_bot/cell_workers/follow_cluster.py
@@ -1,7 +1,7 @@
 from pokemongo_bot.step_walker import StepWalker
 from pokemongo_bot.cell_workers.utils import distance
 from pokemongo_bot.cell_workers.utils import find_biggest_cluster
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 class FollowCluster(BaseTask):
