+++ pokemongo_bot/cell_workers/recycle_items.py
@@ -1,6 +1,6 @@
 import json
 import os
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 from pokemongo_bot.tree_config_builder import ConfigException
 
 class RecycleItems(BaseTask):
