+++ pokemongo_bot/cell_workers/sleep_schedule.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta
 from time import sleep
 from random import uniform
-from pokemongo_bot.cell_workers.base_task import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 
 class SleepSchedule(BaseTask):
