+++ pokemongo_bot/test/resources/plugin_fixture/fake_task.py
@@ -1,4 +1,4 @@
-from pokemongo_bot.cell_workers import BaseTask
+from pokemongo_bot.base_task import BaseTask
 
 class FakeTask(BaseTask):
   def work(self):
